ref: Revert tests
