+++ packages/apm/test/hub.test.ts
@@ -11,6 +11,28 @@ describe('Hub', () => {
     jest.useRealTimers();
   });
 
+  describe('getTransaction', () => {
+    test('simple invoke', () => {
+      const hub = new Hub(new BrowserClient({ tracesSampleRate: 1 }));
+      const transaction = hub.startTransaction({ name: 'foo' });
+      hub.configureScope(scope => {
+        scope.setSpan(transaction);
+      });
+      hub.configureScope(s => {
+        expect(s.getTransaction()).toBe(transaction);
+      });
+    });
+
+    test('not invoke', () => {
+      const hub = new Hub(new BrowserClient({ tracesSampleRate: 1 }));
+      const transaction = hub.startTransaction({ name: 'foo' });
+      hub.configureScope(s => {
+        expect(s.getTransaction()).toBeUndefined();
+      });
+      transaction.finish();
+    });
+  });
+
   describe('spans', () => {
     describe('sampling', () => {
       test('set tracesSampleRate 0 on span', () => {
